Rename Oi*Register to Oi*Bundle
Following a longer discord discussion with Lukas, rename the Oi*Register to Oi*Bundle.
This is because Registers [sic! => registry] are usually filled and (re-)configured at runtime,
while these classes are rather immutable classes that form a functional unit
for e.g. a chain, a contract, a protocol etc.

Further rename related files to bundle.ts and move to top-level of each package/src