core: Purge exported oiCorePlugin w/o changing functionality
For most external cases, the OiCore via async getOiCore() is available.
OiCore got a .getPlugin(namespace, pluginName), which returns fully initialized plugins.
This is more intuitive, since the average developer will not expact to receive
non-initialized plugins.

If retrieval of potentially uninitialized plugins is needed,
e.g. for other plugins, which use it in constructors and their init functions
and are aware the plugins may not be initialized yet,
OiPluginRegistry.getInstance() can be used directly.

This further adds a warning-message, when getOiCore() is called multiple times
in parallel during initialization as well as add a safe-guard to OiCore class
to ensure it can only be initialized once.