execute: Decoded results, support for JSON as function args
Modern smart contracts often have structs and the likes as function args
or return values.

This allows to use JSON-syntax to provide struct-like arguments. Since
uint256 is a very common EVM-datatype, json-bigint is needed as a library
as the native JSON.stringify / JSON.parse does not support BigInt.

This also adds a decoder which can decode nested data types according
to the ABI-definition